Move header files to the interface/ directory
Move header files referenced from outside their directory to the interface/ directory.
Update the #include guards accordingly.
